Add optional callback to onYes() in Modal.js
I attempted a Promise strategy, but it complicated the interface for the
user. Now, Modal.js intelligently determines if the onYes defined by the
user has an arity of 1 or 0. If its arity is 1, then it assumes it's a
callback and closes the modal depending on whether the callback returns
a Boolean value of true.